+++ src/components/Pokestats/elements/index.ts
@@ -1,3 +1,4 @@
 export { default as DisplaySettings } from './DisplaySettings/DisplaySettings';
 export { default as Image } from './Image/Image';
 export { default as Summary } from './Summary/Summary';
+export { default as Tip } from './Tip/Tip';