+++ src/constants/types.ts
@@ -56,6 +56,7 @@ export interface AppState {
   editRule: (newRule: TRuleContent, i: number) => void;
   exportTeamMember: (detail: PokemonDetail) => void;
   importState: (newAppState: Partial<AppState>) => void;
+  massImport: (encounters: TEncounter[]) => void;
   removeNew: () => void;
   reorderRule: (destinationId: number, rule: TRuleEntry, sourceId: number) => void;
   resetAll: () => void;
@@ -292,3 +293,9 @@ export type TEffectiveness = {
 export type TGenerationEffects = {
   [key: string]: TEffectiveness;
 };
+
+export interface ParseResult<T> {
+  data: T[];
+  errors: unknown[];
+  meta: unknown;
+}